@@ -2,19 +2,12 @@
   "name": "multi-step-form-main",
   "version": "1.0.1",
   "private": true,
+  "type": "module",
   "scripts": {
-    "build": "env-cmd -f env/.env.development next build",
-    "start": "env-cmd -f env/.env.development next start -p 3000",
     "dev": "env-cmd -f env/.env.local next dev -p 3000",
-    "build:local": "env-cmd -f env/.env.local next build && next export",
-    "start:local": "env-cmd -f env/.env.local next start -p 3000",
-    "build:dev": "env-cmd -f env/.env.development next build && next export",
-    "start:dev": "env-cmd -f env/.env.development next start -p 3000",
-    "build:prod": "env-cmd -f env/.env.production next build && next export",
-    "start:prod": "env-cmd -f env/.env.production next start -p 3000",
-    "server:start": "node server.js",
-    "lint": "eslint .",
-    "test": "jest --coverage"
+    "build": "env-cmd -f env/.env.production next build",
+    "start": "env-cmd -f env/.env.production next start -p 3000",
+    "lint": "eslint ./src"
   },
   "description": "",
   "main": ".eslintrc.js",
@@ -31,16 +24,11 @@
     "@reduxjs/toolkit": "^2.8.2",
     "@tanstack/react-query": "^5.76.1",
     "@types/cookie": "^0.6.0",
-    "@types/next-pwa": "^5.6.9",
-    "@types/node": "^22.15.21",
-    "@types/react-datepicker": "^6.2.0",
-    "@types/react-redux": "^7.1.34",
     "@types/styled-components": "^5.1.34",
     "antd": "^5.25.4",
     "axios": "^1.9.0",
     "cookie": "^1.0.2",
-    "env-cmd": "^10.1.0",
-    "i18next": "^23.7.0",
+    "i18next": "^25.2.0",
     "next": "^15.3.2",
     "next-pwa": "^5.6.0",
     "next-redux-wrapper": "^8.1.0",
@@ -48,41 +36,29 @@
     "react-datepicker": "^8.3.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.57.0",
-    "react-i18next": "^14.0.0",
+    "react-i18next": "^15.5.1",
     "react-imask": "^7.6.1",
     "react-loader-spinner": "^6.1.6",
     "react-redux": "^9.2.0",
-    "react-scripts": "^5.0.1",
-    "redux": "^5.0.1",
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
-    "redux-thunk": "^3.1.0",
     "styled-components": "^6.1.18",
     "zod": "^3.25.51"
   },
   "devDependencies": {
-    "@eslint/js": "^9.28.0",
+    "@eslint/js": "^9.29.0",
     "@redux-devtools/extension": "^3.3.0",
-    "@tanstack/eslint-plugin-query": "^5.74.7",
-    "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^16.3.0",
-    "@types/jest": "^29.5.14",
     "@types/react": "^19.1.6",
     "@types/react-dom": "^19.1.5",
-    "@types/redux-logger": "^3.0.13",
-    "@types/redux-persist": "^4.0.0",
     "@typescript-eslint/eslint-plugin": "^8.34.0",
     "@typescript-eslint/parser": "^8.34.0",
     "babel-plugin-styled-components": "^2.1.4",
-    "babel-plugin-transform-remove-console": "^6.9.4",
-    "eslint": "^9.28.0",
+    "env-cmd": "^10.1.0",
+    "eslint": "^9.29.0",
     "eslint-config-prettier": "^10.1.5",
     "eslint-plugin-prettier": "^5.4.0",
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^5.2.0",
-    "globals": "^16.2.0",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.3.4",
-    "typescript": "^4.9.5"
+    "typescript": "^5.4.5"
   }
 }
